Add tox test-runner, and very basic tests
This test suite exists so that we can verify the success of a pull request
submitted to this package that supposedly gives us Python 3.x support.

We would do better to have a more serious test suite, but this is
noticeably better than nothing, as it also provides a tox.ini which
will enable us to run the test suite on multiple versions of Python
easily.